Simple form

+++ src/main/java/com/bernardomg/tabletop/palette/palette/model/persistence/PaletteGroup.java
@@ -54,7 +54,7 @@ public class PaletteGroup implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false, unique = true)
-    private Integer           id               = -1;
+    private Long              id               = -1l;
 
     /**
      * Name of the entity.
@@ -98,7 +98,7 @@ public class PaletteGroup implements Serializable {
      *
      * @return the entity's identifier
      */
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
@@ -111,7 +111,7 @@ public class PaletteGroup implements Serializable {
         return Objects.hash(id);
     }
 
-    public void setId(final Integer value) {
+    public void setId(final Long value) {
         id = checkNotNull(value, "Received a null pointer as identifier");
     }
